Correct binary operations to fill d2d.n with int16 instead of double; found this error also in fit.m for d4d; checked all the other operations (func_eval, cut, smooth ...) and these all are OK.
fixes #19


git-svn-id: https://svn.isis.rl.ac.uk/Horace/trunk@120 1cb955da-4463-431d-88dd-2af2f5550de3

+++ @d4d/fit.m
@@ -95,5 +95,5 @@ for i = 1:length(win)
     
     wout(i).s = reshape(sout,size(win(i).s));
     wout(i).e = zeros(size(win(i).e));  
-    wout(i).n = double(~isnan(wout(i).s));
+    wout(i).n = int16(~isnan(wout(i).s));
 end
